Move test context creator to utils
